requirements.txt
beautifulsoup4>=4.9.0
requests>=2.25.0
python-dotenv>=0.19.0
lxml>=4.6.0
sqlalchemy>=1.4.0
python-dateutil>=2.8.0

